make_picks.py: Starting script
This is the first version of the script. First, we look for the
settings.json file, and throw an error if it's not found. After
launching the browser and logging in, we wait for the page to load and
begin parsing each matchup. If the game has not yet been picked, the
home team will be selected. Finally, if the MNF tiebreaker has not yet
been set, we set it to 44.